feat: move alert trigger create in dedicated package
